requirements.txt
flask
flask_cors
hypercorn
openai
python-dotenv
pydantic
